+++ prediction_models/XGBoost.py
@@ -34,7 +34,7 @@ class XGBoostSlim(PredictionModel):
 		if(test):
 			predictions = xgb.predict(X_test)
 			self.model_output['mean_absolute_error'] = mean_absolute_error(y_test, predictions)
-			self.model_output['root_mean_squared_error'] = np.sqrt(mean_squared_error(y_test, predictions))
+			self.model_output['root_mean_squared_error'] = float(np.sqrt(mean_squared_error(y_test, predictions)))
 			importance = pd.DataFrame({
 				'feature': self.team_specific_feature_columns,
 				'importance': xgb.feature_importances_
